Update patch set 2
Patch Set 2:

It was not that in the end, Ethan. The problem came from the fact that the manifest I was building included the Omni's TWRP instead of your recovery, and it was fine on 12.0. However, in 12.1 you introduced the Transparent compression support, which needed few other libraries of lz4 to be imported, and were not added as of now to TWRP. So I switched back to CM's recovery sources and it's all fine now.

Patch-set: 2
Label: Code-Review=0