Update patch set 2
Patch Set 2:

This won't fix any build issues.

That said, I don't have a problem with this if you guys really want it built.  But it's not necessary at all.  We are only using the host library in make_ext4fs, and most or all distros should have a functional lz4 package available.

Patch-set: 2