Update patch set 2
Patch Set 2:

Unlike Martin, I don't see the problem building lz4c too but either way this is actually nice.  To remove a dependency on an external package that has distro-to-distro differences, this is all I had to do:

http://review.cyanogenmod.org/#/c/101080/1

(well had to do and at the same time cleaned up the crap that i had to add to make it build on my build server and jenkins)

I would love to see this merged either with or without the lz4c command line.  Hopefully the two of you can agree on it.

Patch-set: 2
Label: Code-Review=0